Move VSPRtree to index package
